third_party/WebKit/LayoutTests/bluetooth/gattserverdisconnected-event/disconnected.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
 'use strict';
 promise_test(t => {
   return setBluetoothFakeAdapter('DisconnectingHeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: [request_disconnection_service_uuid]
    }))
    .then(device => {
      return device.gatt.connect()
        .then(gattServer => gattServer.getPrimaryService(request_disconnection_service_uuid))
        .then(service => service.getCharacteristic(request_disconnection_characteristic_uuid))
        .then(requestDisconnection => {
          // This promise hangs unless gattserverdisconnected is fired.
          let disconnected = eventPromise(device, 'gattserverdisconnected');
          return Promise.all([requestDisconnection.writeValue(new Uint8Array([0])),
                              disconnected]);
        }).then(([, disconnected]) => assert_true(disconnected.bubbles));
    });
 }, 'A device disconnecting while connected should fire the gattserverdisconnected event.');
</script>
